[bug #165645049] this commit fixes validation bug on view specific account details by user or staff, rearranges the files and fixes some status codes
