+++ .eslintrc.json
@@ -27,6 +27,7 @@
         "no-undef": 0,
         " no-self-compare": 0,
         "import/first":0,
-        "array-callback-return": 0
+        "array-callback-return": 0,
+        "no-else-return": 0
     }
 }
