+++ server/middleware/auth.js
@@ -18,6 +18,7 @@ class Authorization {
         });
       } else {
         req.user = decode;
+        console.log(decode.id);
         next();
       }
     });
